feat(nuxt): autocomplete layouts in `setPageLayout`/`<NuxtLayout>` (#22362)
